Completed module on Component
It is easy to follow the video along, however the exercises
show that it is easy to make simple mistakes and for application
to break. Case in point: not referencing the path to HTML templates
correctly. It does not help that the error messages in Plunker are
useless.

Working extensively with Angular and VSCode also highlighted that
some settings are incorrect, such as deleting whitespaces after a
delay of 1 second! This is corrected now.

Also installed a few extensions to speed-up coding. Angular snippets
are particularly useful.

   --- Dr. Lionel Saliou, Ph.D