Lazy loading of the user component
Observe how the component was not imported in the main
app.module.ts file. As per demo in the video, the code for
user component will only be downloaded when the user navigates
to the relevant link. Furthermore, if there are errors... like I
did ;-) then these will only occur when the component is loaded.

   --- Dr. Lionel Saliou, Ph.D