Created barrels for easier import
When using this technique, it is important to also revisit
parts of the applications that import or reference a lot of
the components, such as the routes!

In this commit, we had to change the name of one of the file as it
was not within the standard. As this file was already badly referenced
elsewhere, the application was failing to compile. It took a couple
of minutes to locate what the error was. Still, simplifying the
imports using barrels at this stage made quick work for ths problem.

   --- Dr. Lionel Saliou, Ph.D