Implemented Template based validation
As part of this exercise, also completed the challenge associated
with wiring a form to call a service. This should have been
completed previously. Noted that the following exercise, which like
this commit focuses on validation, does not bother with extracting
each form fields and place them in the correct "object" location.
Rather the form is constructed in such a way that the form.value
produces the object of the correct shape. This is perhaps
because they use a concept no yet taught of `ngModelFormGroup`.

Whilst, this commit looks "easy" the associated challenge is an
exercise in tediousness. As per the authors' own comment, doing
extensive validation is bothersome in this approach. I quite agree.

    --- Lionel Saliou, Ph.D