Extended create event using template form
I had assumed that this part of the application would leverage
reactive forms. I am a bit concerned by the objective to eventually
use 'directive' in order to perform cross field validations
in template based form... knowing that this is trickier than
re-active forms.

It was an interesting shortcut to bind the ngModel within the HTML
with fields from an object, thus avoiding cluter in the component
TypeScript file itself.

    --- Lionel Saliou, Ph.D