Added Create Session using Reactive form
In this commit, we solved the problem with regards to ordering
within the barrel. There were issues because we imported models
AFTER defining services that used them.

In this commit, we are still using Angular built-in validators.

There appear to be a problem with the maxLength whereby the
text area turns red when 400 characters or more are typed in,
however the corresponding error message does not appear above
the text area. The "required" one appears just fine.

   --- Lionel Saliou, Ph.D