Implemented shared custom reactive form validator
Encountered a weird problem with the file not being found. Perhaps,
the dash in VSCode is not the same as the dash in the Git Console
for example (UTF8 vs ASCII?).

It is a bit tricky to understand the fat arrow syntax at time.
Likewise, the function chaining needs careful consideration.

I have not been able to see the Max Length error message yet.
I note that this was not demonstrated in the course either.

    --- Lionel Saliou, Ph.D