FIXED compile error on Lazy module 1st serve
During this chapter development, it was noticed that the application
would not start the first time it was served. There was an error
pointing at the fact that the `routes.ts` file could not access
or locate the user.module. Oddly enough, modifying the user.module
file whilst `ng serve` was running would solve this problem.

Searching online showed that this is a problem for the community.
In order to sort this out, one must use the relative path notation
in routes.ts combined with the `baseUrl:` paramter in the compiler
options.

    --- Lionel Saliou, Ph.D