FIXED length validation problem
FINALLY! I managed to fix the display of the error message when
the session's abstract contains more than 400 characters.

There are several posts online that suggests that this feature
is not working. Honnestly, I should have read the documentation
in the first place. Indeed, it clearly states that the field
within the error object is spelled 'maxlength' (observe the
capatilisation) whereas we used 'maxLength' in the HTML template.

As noted in a previous commit, the assumption was that the name
of the key within the error object was the same as the name of
the validator. It is the case for the standard 'required' validator
for example. The Angular team has made a poor job of their naming
covention here. Arguably, they could have specified the error object
type (Note: this would have been helpful in the HTML template as
it is not yet possible to take advantage of the strong typing
there).

    --- Lionel Saliou, Ph.D