Handle data emitted from child component
In this commit, we show how to pass data back from the child
component back to the parent. For example, we pass the newly
form ISession object up the chain using the @Output decorator.
Thus, in the HTML we bind the parent to this property and when
it is triggered, the parent is able to extract the data from
the `$event` object. In our case, this data is redirected to
the corresponding 'service' functionality.

Likewise, we use the same technique to toggle from the form. When
the child component hits "close", the parent reacts by setting
the 'addMode' back to fault. This in turn forces Angular to run
the DOM change detection and the form disappears.

The only reproach I would make is that it can be confusing to
read the code as often time we see the same names used both side
of the equality sign. Perhaps, the parent side should be called
eventNameHandler at the very least so that the corresponding
function in the TS file can be uniquely identified.

The handlers can be defined such that the parameter object can be
strongly typed. However, in this commit, this was not tested. In other
words, we did not try to submit an object of the wrong "shape" to
see what would happen.

    --- Lionel Saliou, Ph.D