Implemented session filtering
In this commit, we took the following approach:

1. Buttons in the parent component define the filter value for
   the component.
2. The filter value is passed down to the child component.
3. The child component in turns interprets the filter only to
   return a "sub-set" (possibly) of all of the sessions it was given.

It is important to note that we could have implemented the filtering
on the parent component so it would prepare the data for the child
component.

The filtering approach would have remained the same. However the manner
in which the various array would be defined or changed would likely
be different.

    --- Lionel Saliou, Ph.D