Implemented session sorting
As with previous commit, we could have done that in the parent
component, but still in this instance with supply the ordering
choice down to the child component.

Notably, we had to define sorting functions. These have nothing
to do with the component itself, thus their logic was defined
elsewhere.

Initially, these were defined within the event.model file. The
application would compile but fail to run as the compiler could
not find the export for these function in the barrel file (index.ts).

Restarting the application sorted the issue.

   --- Lionel Saliou, Ph.D