Rework the Toastr service to use Opaque Token
In this commit, we learn how to better wrapper a 3rd party element
within an Angular 2+ application so that it better fits within the
dependency injection model. The purpose of this approach/technique
is to limit the possibility of collisions. This is a great risk
when strings are used to lookup elements. Using objects, such
as it is done here, tremendously limits this risk and allows
for using singletons.

NOTE/BUG: in a previous commit, we "fixed" a build issue as with
the Lazy user model. Sadly, as part of this commit we forgot to
test this route. Testing the TOASTR_TOKEN showed that whilst
the build was successful the application was broken. The Login
page DOES NOT appear.

    --- Lionel Saliou, Ph.D