Fix Lazy Login Component loading
Without this change, the Login page does not appear. There are
errors in the console related to the fact the component cannot be
found.

This change corrects this issue but it is likely to "re-open" the
problem whereby the application will not correctly build the
first time we call `ng serve`.

     --- Lionel Saliou, Ph.D