Using jQuery to create a modal comp & directive
This is quite a large commit.

1. We create an area within the Nav-bar component where matching
results can be displayed.
2. We delegate the search to the Event Service.
3. The event service returns an Observable.
4. The Nav Bar component subscribe to the corresponding service
from the Event Service component.
5. The Nav-bar component writes its results into a component
6. This component is used for projection so that the content
can be formatted in such a way that a modal (jQuery compatible)
list can be shown.
7. In order to hook into the jQuery functionality, we create a
directive on the Search Button. This allows for introducing a level
of indirection, so that the Nav Bar component does not need to know
how to set the modal/non-modal mode. It may not, in the future,
use jQuery at all, for example.

BUG: Once, we have navigated to our chosen result, chosing another
session does not redirect to an event. Looking at the console,
it appears as if the routerLink data is correct.

    --- Lionel Saliou, Ph.D