Improve modal component flexibility
Very common in Angular, fixed values can cause problem. Prior to
this commit, it was not possible to define any additional modal
components because the ID used as part of the jQuery call was
hard coded. Likewise, the directive applied to only 1 specific
element. This made the component non-reusable.

In this commit, we demonstrate how to parameterise the component
as well as the directive so that multiple such components can
exists in 1 application. Here, we could have a list of user that
is displayed in the same manner.

   --- Lionel Saliou, Ph.D