Fixed event details navigating to self
This issue was due to the fact that we used the snapshots from
the route. This value was fixed when the component is first
instanciated. Even though through the list we wanted to navigate
the component state was not changing.

With this commit, we iterate through each of the parameters
passed to the component and change the component states accordingly.
It is "states" plural because the component has more than 1, such
as the addMode. Arguably, the filters could be considered states
too.

    --- Lionel Saliou, Ph.D