Implement ability to close model on click
This feature uses the ViewChild decorator. There are many other
similar decoractor, however the instructors believe their use
is quite rare as they are quite complex.

Meanwhile, the key here is to understand that we use public
reference variable defined in the HTML (eg #varName) as a mean
to define this ViewChild and we cast it to a ElementRef object.
Thus we are able to access the underlyung DOM element and then
call a jQuery method on it.

    --- Lionel Saliou, Ph.D