Implement UpVote component with associated service
This was quite a bit of work. We created a new fairly complex
visual component which takes in various parameter and emit an
event. Along with this, by communicating up to its parent, it may
change the way the parent renders the session list.

It has become easier to keep track of the properties of component
across the HTML and TS elements.

   --- Lionel Saliou, Ph.D