Implement a complex form validator
This was quite a bit of work. A lot of moving part!

This is a pattern that would have to be re-visited if a future
project needs something along those lines.

Perhaps, this is due to the nature of the form itself. After all,
the create event form is a template based form. A reactive form
might be easier to validate.

    --- Lionel Saliou, Ph.D