Adding ngf-server dependency for data over HTTP
This is a server side application that the author wrote that
allows for serving data over HTTP and thus simulate how a real or
production Angular application would interact with a datasource.

EDIT: also added a new command to the `package.json` file so that
this application above can be started from the command-line.

   --- Lionel Saliou, Ph.D