Define a proxy file for dev mode
Basically, this file is used in development mode. It tells the
CLI server (which is responsible for hosting our Angular app
whilst in local development) that when a request comes in for a
URL containing "/api" this needs to be sent to the ngf-server.

EDIT: changed the `start` command so that the proxy configuration
is loaded.

NOTE:
1. to start the HTTP Server, use a command prompt and use
`npm run server`
2. to start the App using the proxy, use the command prompt:
`npm start`

IDEA:
need to create a `dev.ps1` script so that both can be started
correctly whilst in development mode.

     --- Lionel Saliou, Ph.D