Simply gathering data from the resolver
This commit is to highlight the fact that the 'data' obtained
through the resolver is also assigned to an array on the route
object itself. Thus, instead of looping through the parameters of
route (which might not be necessary), then we copy the gathered
data directly from the route.data field.

    ---- Lionel Saliou, Ph.D