Update service to allow updating data via HTTP
A worthy item here: the author elected to code in the fact that
the POST request might return an object of IEvent type, so that
the client (for instance) could retrieved the Event ID as it is
a mandatory piece of information that the client cannot choose.

That said, the function that handles the response for this request
does not care and thus defines the fat arrow lambda using the
() => notation.

    ---- Lionel Saliou, Ph.D