Replace updateEvent with saveEvent
This is due to the fact that authors have implemented smart end-
point. Consequently, the server knows that when an event is supplied
with now Event ID, it is a new record to be created (hence using the
verb POST). Conversely, when the Event ID is supplied, the server
knows it is an update where the client side should really be using
a PUT verb. Because the smart end-point, we do not have to do this.

    --- Lionel Saliu, Ph.D