Change output directory to serve production build
The authors had us change the `npm run server` command so that
their ngf-server would be able to both be used as an API end-point
as well as provide whatever is placed in the `dist/` directory.

For evidence of that, see the `expressConfig.js` file at line 25.
It is clear from the code that the node server is expecting to
find something two folders above.

This change allows this application to be compatible with their
expectations.

    --- Lionel Saliou, Ph.D