Fix use of Toastr in production mode
Using the tutorial from the Toastr author, combined with the
the level of indirection that is offered through the existing
"ToastrService", it was fairly easy to swap the implementation.
Along with this, I also included a custom "option" object
for Toastr. It appears to be working just fine.

I wish I could say the same about jQuery :'(

     --- Lionel Saliou, Ph.D

EDIT: navbar component fix. As the authors of the course hinted
it is possible to access private component properties in the
template in development mode. This is not allowed in production.