Fix jQuery/bootstrap modal in Production mode
Finally managed to get the modal form to show / be usable
whilst running the application in Production mode.

It is fair to say that I did not interpret the error messages
correctly. The compiler seems to be very forgiving when developping
actively.

What I need to remember is that all of the symbols I might require
need to be encapsulated in their own files and/or component/object.
I can only guess that it is to ensure that should other component
reference these, the compiler has the opportunity to keep consistency
across the application.

It would have been better if there were warnings during development
because it is a lot of effort to fix these problems for beginners
like myself.

*IMPORTANT NOTE*: This issue was not strictly a jQuery problem. Yes,
there were issues with ensuring that our application was capable
of using this framework. However, the `modal` property is from
bootstrap. Bootstrap itself has a dependency on jQuery (see the
`package` files). By ensuring that we were giving the components
access to jQuery correctly, we re-established the application
capabilities to properly use Bootstrap.

   --- Lionel Saliou, Ph.D