Document getting started on Windows 10
It was important to revisit how to clone and get the environment
working in MS Windows 10.

Two tasks ahead of us:
1. Creating containers for this work:
a) One for the ngf-server that contains the data and login capabilities
b) One for the application itself.

It must be noted that the CLI version is now 6.0.8 whereas the
package JSON originally stated 6.0.3.

Also, the Lock file is changed such that the versions may change
very soon, whereas the course aimed to keep the version fixed.

Once both elements of the required work are containerised, we will
attempt to address the issues GitHub is highlighting.

    --- Lionel Saliou, Ph.D