Dockerise application with docker-compose
With this change, ng-fundamental becomes easier to deploy and use
by relying on the `docker-compose` executable. A set of containers
are defined to encapsulate the required proxy that fronts the
application itself (this repo) as well as the API server it
connects to (ngf-server). This made possible by the docker engine
as each container is able to look-up others through internal DNS.

Environmental variables are used to provide some flexibility and
ease of maintenance. Along with this, this appraoch allows us to
quickly refer or use specific version of the containers without
changing the `docker-compose.yml` file. We could override the
required values using the `-e` switch of the `docker-compose`
executable.

The NGNIX configuration for 2 of the containers is simple and
is likely to have room to improve on performance and/or stability.

This commit offers a stable basis to investigate how to resolve
the issues flag during the build by `npm` regarding modules with
various security problems.

Finally, there is a need to investigate how to integrate this command:
``docker images --filter "dangling=true" -q`` so that all
intermediate images are deleted from time to time.

    --- Lionel Saliou, Ph.D